cleanup demo structure
